+++ src/collective/documentgenerator/events/styles_events.py
@@ -48,6 +48,14 @@ def podtemplate_modified(pod_template, event):
     update_PODtemplate_styles(pod_template, event)
 
 
+def styletemplate_created(style_template, event):
+    """
+    Set the md5 of the initial style template in 'initial_md5' field.
+    """
+    if not style_template.initial_md5:
+        style_template.initial_md5 = style_template.current_md5
+
+
 def update_PODtemplate_styles(pod_template, event):
     """
     Update styles on a pod_template using external styles.
@@ -64,6 +72,9 @@ def _update_template_styles(pod_template, style_template_filename):
     """
     Update template pod_template by templateStyle.
     """
+    # we check if the pod_template has been modified except by style only
+    style_changes_only = pod_template.style_modification_md5 and \
+        pod_template.current_md5 == pod_template.style_modification_md5
     # save in temporary file, the template
     temp_file = create_temporary_file(pod_template.odt_file, 'pod_template.odt')
     new_template = open(temp_file.name, 'w')
@@ -81,11 +92,17 @@ def _update_template_styles(pod_template, style_template_filename):
     )
     (stdout, stderr) = executeCommand(cmd.split())
     if stderr:
-        request = pod_template.REQUEST
-        api.portal.show_message(message=_(u"Problem during styles update on template '${tmpl}': ${err}",
-                                mapping={'tmpl': safe_unicode(pod_template.absolute_url_path()),
-                                         'err': safe_unicode(stderr)}),
-                                request=request, type='error')
+        logger.error("Error during command '%s'" % cmd)
+        logger.error("Error is '%s'" % stderr)
+        portal = api.portal.get()
+        request = portal.REQUEST
+        try:
+            api.portal.show_message(message=_(u"Problem during styles update on template '${tmpl}': ${err}",
+                                    mapping={'tmpl': safe_unicode(pod_template.absolute_url_path()),
+                                             'err': safe_unicode(stderr)}),
+                                    request=request, type='error')
+        except:
+            pass
         raise Redirect(request.get('ACTUAL_URL'), _(u"Problem during styles update on template '${tmpl}': ${err}",
                                                     mapping={'tmpl': safe_unicode(pod_template.absolute_url_path()),
                                                              'err': safe_unicode(stderr)}))
@@ -100,6 +117,9 @@ def _update_template_styles(pod_template, style_template_filename):
                                filename=pod_template.odt_file.filename)
         pod_template.odt_file = result
         os.remove(resTempFileName)
+        #if only styles were modified: update the style_modification_md5 attribute
+        if style_changes_only:
+            pod_template.style_modification_md5 = pod_template.current_md5
 
     os.remove(temp_file.name)